Refactor webhook handling by introducing a new Webhook service and removing the old handler. Update main application to utilize the new service for wallet webhooks. Clean up configuration by removing unused WebhookURL variable.
