+++ internal/v3/handler/handler.go
@@ -26,6 +26,7 @@ import (
 	"github.com/ton-connect/bridge/internal/models"
 	"github.com/ton-connect/bridge/internal/ntp"
 	"github.com/ton-connect/bridge/internal/utils"
+	"github.com/ton-connect/bridge/internal/webhook"
 	storagev3 "github.com/ton-connect/bridge/internal/v3/storage"
 )
 
@@ -74,10 +75,10 @@ type handler struct {
 	realIP            *utils.RealIPExtractor
 	eventCollector    analytics.EventCollector
 	eventBuilder      analytics.EventBuilder
-	walletWebhook     *handler_common.WalletWebhookService
+	walletWebhook     *webhook.Service
 }
 
-func NewHandler(s storagev3.Storage, heartbeatInterval time.Duration, extractor *utils.RealIPExtractor, timeProvider ntp.TimeProvider, collector analytics.EventCollector, builder analytics.EventBuilder, walletWebhook *handler_common.WalletWebhookService) *handler {
+func NewHandler(s storagev3.Storage, heartbeatInterval time.Duration, extractor *utils.RealIPExtractor, timeProvider ntp.TimeProvider, collector analytics.EventCollector, builder analytics.EventBuilder, walletWebhook *webhook.Service) *handler {
 	// TODO support extractor in v3
 	h := handler{
 		Mux:               sync.RWMutex{},
@@ -361,16 +362,13 @@ func (h *handler) SendMessageHandler(c echo.Context) error {
 	topic := ""
 	if ok {
 		topic = topicParam[0]
-		go func(clientID, topic, message string) {
-			handler_common.SendWebhook(clientID, handler_common.WebhookData{Topic: topic, Hash: message})
-		}(clientID.String(), topic, string(message))
 	}
 
 	if h.walletWebhook != nil {
 		if walletParam, ok := params["wallet"]; ok && len(walletParam) > 0 {
 			wallet := walletParam[0]
 			if webhookURL, ok := h.walletWebhook.GetWebhookURL(wallet); ok {
-				go h.walletWebhook.Send(webhookURL, handler_common.WalletWebhookData{
+				go h.walletWebhook.Send(webhookURL, webhook.Data{
 					ClientID: clientID.String(),
 					To:       toId.String(),
 					Message:  string(message),
